Implemeted loading states on the forgot and reste password pages
